+++ tests/interposer_test.py
@@ -6,17 +6,22 @@
 import logging
 import shutil
 import tempfile
+import types
 import unittest
 import uuid
 
+from contextlib import contextmanager
+from copy import deepcopy
 from datetime import datetime
 from enum import Enum
 from pathlib import Path
 from typing import Dict
+from typing import Generator
 
 from interposer import Interposer
 from interposer import Mode
 from interposer import PlaybackError
+from interposer import ResultHandlingFlag
 from interposer import ScopedInterposer
 from interposer.interposer import _InterposerWrapper
 
@@ -41,10 +46,17 @@ class MyVerySpecificError(RuntimeError):
     pass
 
 
+def numberGenerator(n):
+    if n < 20:
+        number = 0
+        while number < n:
+            yield number
+            number += 1
+    else:
+        return
+
+
 class SomeClass(object):
-
-    throw_exception = True
-
     def __init__(self, result: object, secret: str = None):
         """
         Store a result for say_hi.  If this derives from Exception then
@@ -52,6 +64,9 @@ class SomeClass(object):
         """
         self.result = result
         self.auth = {"secret": secret}
+
+        self.throw_exception = True
+        self.throw_if_range_called = False
 
     def say_hi(self, greeting: str = "hello", second: object = None) -> str:
         """
@@ -67,6 +82,16 @@ class SomeClass(object):
     def give_up(self):
         if self.throw_exception:
             raise MyVerySpecificError("ouchies")
+
+    def generated_numbers(self, n: int) -> Generator[int, None, None]:
+        """
+        Returns a generator.  These are harder to deal with
+        because they only generate their contents once, so we
+        have to extract the contents before we record and return.
+        """
+        if self.throw_if_range_called:
+            raise NotImplementedError("should have satisfied result from recording!")
+        return numberGenerator(5)
 
     @property
     def get_complex_stuff(self):
@@ -86,22 +111,76 @@ class SomeClassSecretRemoverInterposer(Interposer):
     2. By removing it from any result.
     """
 
-    def cleanup_parameters_pre(self, params) -> Dict:
+    def cleanup_parameters_pre(self, params: Dict) -> Dict:
         """
         Remove the secret from the parameters used to initialize it
         in the recording.
         """
         if "secret" in params["kwargs"]:
-            params["kwargs"]["secret"] = "REDACTED_SECRET"  # nosec
-        return params
+            result = deepcopy(params)
+            result["kwargs"]["secret"] = "REDACTED_SECRET"  # nosec
+            return super().cleanup_parameters_pre(result)
+        else:
+            return params
 
-    def cleanup_result_pre(self, params, result) -> object:
+    @contextmanager
+    def cleanup_result_pre(
+        self, params: Dict, result: object, flags: ResultHandlingFlag
+    ) -> None:
         """
         Remove the secret from the initialized class in the recording.
         """
         if "secret" in params["kwargs"]:
+            original_secret = result.__dict__["auth"]["secret"]
             result.__dict__["auth"]["secret"] = "REDACTED_SECRET"  # nosec
-        return result
+            # this is good form in case there are stacked interposers
+            with super().cleanup_result_pre(params, result, flags=flags) as (
+                scrubbed_result,
+                flags,
+            ):
+                yield scrubbed_result, flags
+            result.__dict__["auth"]["secret"] = original_secret
+        else:
+            yield result, flags
+
+
+class SomeClassGeneratorInterposer(Interposer):
+    """
+    Tests handling generators by overwriting the original result.
+    """
+
+    @contextmanager
+    def cleanup_result_pre(
+        self, params: Dict, result: object, flags: ResultHandlingFlag
+    ) -> None:
+        if params["method"] == "generated_numbers":
+            assert isinstance(result, types.GeneratorType)
+            new_result = list(result)
+            flags |= ResultHandlingFlag.REPLACE
+            # this is good form in case there are stacked interposers
+            with super().cleanup_result_pre(params, new_result, flags=flags) as (
+                scrubbed_result,
+                flags,
+            ):
+                yield scrubbed_result, flags
+
+
+class SomeClassRecordNothingInterposer(Interposer):
+    """
+    Not very practical but here we simply state that no recording should be done.
+    """
+
+    @contextmanager
+    def cleanup_result_pre(
+        self, params: Dict, result: object, flags: ResultHandlingFlag
+    ) -> None:
+        flags ^= ResultHandlingFlag.RECORD
+        # this is good form in case there are stacked interposers
+        with super().cleanup_result_pre(params, result, flags=flags) as (
+            scrubbed_result,
+            flags,
+        ):
+            yield scrubbed_result, flags
 
 
 class InterposerTest(unittest.TestCase):
@@ -110,7 +189,6 @@ class InterposerTest(unittest.TestCase):
 
     def setUp(self):
         self.datadir = Path(tempfile.mkdtemp())
-        SomeClass.throw_exception = True
 
     def tearDown(self):
         shutil.rmtree(str(self.datadir))
@@ -163,6 +241,27 @@ class InterposerTest(unittest.TestCase):
             assert wt.say_hi(second=MyEnum.FOO) == "hello True MyEnum.FOO"
             with self.assertRaises(PlaybackError):
                 wt.say_hi(second="foobar")  # never called during recording
+            wt.throw_if_range_called = True
+
+    def test_generators(self):
+        """
+        Tests how we handle generators.
+        """
+        uut = SomeClassGeneratorInterposer(self.datadir / "recording", Mode.Recording)
+        uut.open()
+        wt = uut.wrap(SomeClass(True))
+        # although it would normally return a generator, in order to record
+        # it we drain the generator to a list, record it, and return the list
+        # based on how SomeClassGeneratorInterposer is configured
+        assert wt.generated_numbers(5) == [0, 1, 2, 3, 4]
+        uut.close()
+
+        with ScopedInterposer(self.datadir / "recording", Mode.Playback) as uut:
+            wt = uut.wrap(SomeClass(False))
+            wt.throw_if_range_called = True
+            # if the method is actually called it will now throw
+            # instead we prove it is satisfying the result from the recording
+            assert wt.generated_numbers(5) == [0, 1, 2, 3, 4]
 
     def test_wrappable(self):
         """
@@ -219,7 +318,7 @@ class InterposerTest(unittest.TestCase):
             with self.assertRaises(MyVerySpecificError) as re:
                 wt.give_up()
             self.assertRegex("ouchies", str(re.exception))
-            SomeClass.throw_exception = False
+            wt.throw_exception = False
             self.assertIsNone(wt.give_up())
 
         with ScopedInterposer(self.datadir / "recording", Mode.Playback) as uut:
@@ -229,10 +328,12 @@ class InterposerTest(unittest.TestCase):
             self.assertIn("say_hi", dir(wt))
             # but we are playing back so it returns what was recorded
             self.assertEqual(wt.say_hi(), "hello True")
-            # SomeClass is currently set not to throw, but we're playing back
-            # so we replay the exception
-            with self.assertRaises(MyVerySpecificError) as re:
+            # prove we replay exceptions
+            wt.throw_exception = False  # but the original call threw
+            with self.assertRaises(MyVerySpecificError):
                 wt.give_up()
+            wt.throw_exception = True  # but the original call did not
+            self.assertIsNone(wt.give_up())
 
     def test_property_handling(self):
         """
@@ -330,6 +431,25 @@ class InterposerTest(unittest.TestCase):
                 # was never called a second time
                 wt.say_hi()
 
+    def test_result_flags_disable_recording(self):
+        """
+        Tests that recording flags can stop recording.
+
+        Not intended to be an example, just to exercise the record flag turned off.
+        """
+        uut = SomeClassRecordNothingInterposer(
+            self.datadir / "recording", Mode.Recording
+        )
+        uut.open()
+        wt = uut.wrap(SomeClass)
+        wt(True)
+        uut.close()
+
+        with ScopedInterposer(self.datadir / "recording", Mode.Playback) as uut:
+            wt = uut.wrap(SomeClass)
+            with self.assertRaises(PlaybackError):
+                wt(True)
+
     def test_recording_contains_no_secret(self):
         """
         Check that if we properly implement cleanup_parameters_pre, any secret
@@ -342,7 +462,11 @@ class InterposerTest(unittest.TestCase):
         )
         uut.open()
         secret = str(uuid.uuid4())
-        wt = uut.wrap(SomeClass)  # wraps the class definition
+        # wrap the class definition so that we record the __init__ behavior
+        wt = uut.wrap(SomeClass)
+        # by instantiating an object from the wrapped class definition, we
+        # end up storing "t" in the database (pickled), and during playback
+        # when "t" is instantiated is is provided entirely from the recording.
         t = wt(True, secret=secret)
         self.assertEqual(t.say_hi(), "hello True")
         uut.close()
@@ -350,6 +474,17 @@ class InterposerTest(unittest.TestCase):
         with (self.datadir / "recording").open("rb") as fp:
             data = fp.read()
             assert (
-                "REDACTED_SECRET".encode() in data
+                "REDACTED_SECRET".encode() in data  # nosec
             ), "did not find redacted secret in data file"
             assert secret.encode() not in data, "found original secret in data file"
+
+        with ScopedInterposer(self.datadir / "recording", Mode.Playback) as uut2:
+            wt = uut2.wrap(SomeClass)
+            with self.assertRaises(PlaybackError):
+                # due to the current interposer design, we store the argument lists
+                # in the database so we have to scrub away the secret in the params
+                # which means on playback runs we must provide the redacted version
+                # in order to match
+                wt(True, secret=secret)
+            t = wt(True, secret="REDACTED_SECRET")  # nosec
+            assert t.say_hi() == "hello True"